requirements.txt
langgraph>=0.3.0,<1.0
langchain-core>=0.3.0,<0.4
langchain-google-genai>=2.0.0,<3.0
flask>=3.0.0,<4.0
firebase-admin>=6.5.0,<7.0
PyPDF2>=3.0.0,<4.0
easyocr>=1.7.0,<2.0
crawl4ai>=0.6.0,<1.0
beautifulsoup4>=4.12.0,<5.0
python-magic>=0.4.27,<1.0
python-dotenv>=1.0.0,<2.0
google-generativeai
langgraph-checkpoint-sqlite
libmagic